test apply_tags does subscribe to dashamail
