+++ src/users/tests/tags/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+from django.utils import timezone
+
 
 @pytest.fixture
 def user(user):
@@ -8,6 +10,16 @@ def user(user):
     return user
 
 
+@pytest.fixture
+def paid_order(factory, user):
+    return factory.order(user=user, paid=timezone.now(), unpaid=None, shipped=None)
+
+
+@pytest.fixture
+def unpaid_order(factory, user):
+    return factory.order(user=user, unpaid=timezone.now(), paid=None, shipped=None)
+
+
 @pytest.fixture(autouse=True)
-def _mock_subscribe_to_dashamail(mocker):
-    mocker.patch("app.integrations.dashamail.helpers.subscribe_user_to_dashamail")
+def mock_subscribe_to_dashamail(mocker):
+    return mocker.patch("users.tags.pipeline.subscribe_user_to_dashamail")
